requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
distro==1.9.0
faiss-cpu==1.10.0
filelock==3.18.0
frozenlist==1.5.0
fsspec==2025.3.2
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
huggingface-hub==0.30.2
idna==3.10
Jinja2==3.1.6
jiter==0.9.0
joblib==1.4.2
MarkupSafe==3.0.2
mpmath==1.3.0
multidict==6.4.3
networkx==3.4.2
numpy==2.2.4
openai==1.73.0
packaging==24.2
pillow==11.1.0
propcache==0.3.1
PyAudio==0.2.14
pycparser==2.22
pydantic==2.11.3
pydantic_core==2.33.1
python-dotenv==1.1.0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
sentence-transformers==4.0.2
setuptools==78.1.0
simpleaudio==1.0.4
sniffio==1.3.1
sounddevice==0.5.1
sympy==1.13.1
threadpoolctl==3.6.0
tokenizers==0.21.1
torch==2.6.0
tqdm==4.67.1
transformers==4.51.2
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
wavefile==1.6.3
yarl==1.19.0
